+++ otclient/src/framework/net/inputmessage.cpp
@@ -101,14 +101,14 @@ bool InputMessage::decryptRsa(int size)
     return (getU8() == 0x00);
 }
 
-void InputMessage::fillBuffer(uint8 *buffer, uint16 size)
+void InputMessage::fillBuffer(uint8 *buffer, uint32 size)
 {
     checkWrite(m_readPos + size);
     memcpy(m_buffer + m_readPos, buffer, size);
     m_messageSize += size;
 }
 
-void InputMessage::setHeaderSize(uint16 size)
+void InputMessage::setHeaderSize(uint32 size)
 {
     assert(MAX_HEADER_SIZE - size >= 0);
     m_headerPos = MAX_HEADER_SIZE - size;
